CPN-15 Use window.postMessage for cache invalidation
When we need to eagerly flush the cache we use a window.postMessage to allow the settings UI to tell the loading code that the next time it loads it should not use the cache. We store a value in session storage to keep this across page reloads, but then once we've performed a load we load from the cache again.

I had lots of issues working on where because of the LTI redirect when returning from the setting UI it would partially load the new file, but not properly and would appear to clear out the local storage which ended up in the setting still being cached. I'm not sure about the exact race condition, but moving the setting of the local storage later in the page appears to have fixed it.